Move updates to a service and share it; fix web updates
